+++ primefaces/src/main/java/org/primefaces/component/fileupload/FileUploadRenderer.java
@@ -69,6 +69,12 @@ public class FileUploadRenderer extends CoreRenderer {
     public void encodeEnd(FacesContext context, UIComponent component) throws IOException {
         FileUpload fileUpload = (FileUpload) component;
 
+        if (fileUpload.getDropZone() != null) {
+            fileUpload.setMode("advanced");
+            fileUpload.setAuto(true);
+            fileUpload.setDragDropSupport(true);
+        }
+
         encodeMarkup(context, fileUpload);
         encodeScript(context, fileUpload);
     }
@@ -81,15 +87,21 @@ public class FileUploadRenderer extends CoreRenderer {
         if (fileUpload.getMode().equals("advanced")) {
             PrimeApplicationContext pfContext = PrimeApplicationContext.getCurrentInstance(context);
 
-            wb.init("FileUpload", fileUpload);
+            String dropZone = null;
+            if (fileUpload.getDropZone() != null) {
+                dropZone = SearchExpressionFacade.resolveClientIds(context, fileUpload, fileUpload.getDropZone(),
+                        SearchExpressionUtils.SET_RESOLVE_CLIENT_SIDE);
+            }
 
-            wb.attr("dnd", fileUpload.isDragDropSupport(), true)
+            wb.init("FileUpload", fileUpload)
+                    .attr("dnd", fileUpload.isDragDropSupport(), true)
                     .attr("previewWidth", fileUpload.getPreviewWidth(), 80)
                     .attr("sequentialUploads", fileUpload.isSequential(), false)
                     .attr("maxChunkSize", fileUpload.getMaxChunkSize(), 0)
                     .attr("maxRetries", fileUpload.getMaxRetries(), 30)
                     .attr("retryTimeout", fileUpload.getRetryTimeout(), 1000)
                     .attr("resumeContextPath", pfContext.getFileUploadResumeUrl(), null)
+                    .attr("dropZone", dropZone, null)
                     .callback("onAdd", "function(file, callback)", fileUpload.getOnAdd())
                     .callback("oncancel", "function()", fileUpload.getOncancel())
                     .callback("onupload", "function()", fileUpload.getOnupload());
@@ -139,13 +151,16 @@ public class FileUploadRenderer extends CoreRenderer {
         ResponseWriter writer = context.getResponseWriter();
         String clientId = fileUpload.getClientId(context);
         String style = fileUpload.getStyle();
-        String styleClass = fileUpload.getStyleClass();
-        styleClass = styleClass == null ? FileUpload.CONTAINER_CLASS : FileUpload.CONTAINER_CLASS + " " + styleClass;
+        String styleClass = getStyleClassBuilder(context)
+                .add(FileUpload.CONTAINER_CLASS)
+                .add(fileUpload.getStyleClass())
+                .add(fileUpload.getDropZone() != null, FileUpload.WITHDROPZONE_CLASS)
+                .build();
         boolean disabled = fileUpload.isDisabled();
 
         writer.startElement("div", fileUpload);
         writer.writeAttribute("id", clientId, "id");
-        writer.writeAttribute("class", styleClass, styleClass);
+        writer.writeAttribute("class", styleClass, "styleClass");
         if (style != null) {
             writer.writeAttribute("style", style, "style");
         }